requirements.txt
# Social Media Jobs Dependencies
pymongo>=4.0.0
requests>=2.28.0
google-api-python-client>=2.0.0
youtube-transcript-api>=0.6.0
pytz>=2022.1
python-dotenv>=0.19.0
beautifulsoup4>=4.11.0
python-dateutil>=2.8.0
twscrape>=0.0.1
asyncio
